# documentation/VIDEO_EDITOR_OPENAPI_SPEC.yaml
openapi: 3.0.3
info:
  title: StoryCore Video Editor Wizard API
  description: |
    API complète pour l'éditeur vidéo IA de StoryCore.
    
    ## Fonctionnalités
    - **Authentification** : Inscription, connexion, refresh token
    - **Projets** : CRUD, gestion timeline, templates
    - **Médias** : Upload, download, métadonnées
    - **Export** : Jobs asynchrones, multiples formats
    - **IA** : Transcription, TTS, Smart Crop, Traduction
    
    ## Authentification
    Tous les endpoints (sauf auth) nécessitent un token Bearer JWT.
    
    ## Rate Limiting
    - Gratuit : 10 requêtes/minute
    - Pro : 60 requêtes/minute
    - Team : 120 requêtes/minute
  version: 1.0.0
  contact:
    name: StoryCore Support
    email: support@storycore.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000/api/video-editor
    description: Serveur de développement
  - url: https://api.storycore.ai/api/video-editor
    description: Serveur de production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    Error:
      type: object
      properties:
        detail:
          type: string
          description: Message d'erreur détaillé
        code:
          type: string
          description: Code d'erreur
        status:
          type: integer
          description: Code HTTP
    
    # Auth
...
    ProjectCreate:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
          example: "Mon projet vidéo"
        description:
          type: string
          nullable: true
        aspect_ratio:
          type: string
          default: "16:9"
          enum: ["16:9", "9:16", "1:1", "4:5", "4:3", "21:9"]
        resolution:
          type: string
          default: "1920x1080"
        frame_rate:
          type: number
          default: 30.0
    ProjectUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 255
        description:
          type: string
          nullable: true
        aspect_ratio:
          type: string
          enum: ["16:9", "9:16", "1:1", "4:5", "4:3", "21:9"]
        resolution:
          type: string
        frame_rate:
          type: number
    Project:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
          nullable: true
        user_id:
          type: string
          format: uuid
        aspect_ratio:
          type: string
        resolution:
          type: string
        frame_rate:
          type: number
        duration:
          type: number
          default: 0.0
        created_at:
          type: string
          format: date-time
        modified_at:
          type: string
          format: date-time
        thumbnail_path:
          type: string
          nullable: true
    
    # Media
...
paths:
...
  /projects:
    get:
      tags:
        - Projects
      summary: Lister les projets
      description: Retourne tous les projets de l'utilisateur
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: Liste des projets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    post:
      tags:
        - Projects
      summary: Créer un projet
      description: Crée un nouveau projet vidéo
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectCreate'
      responses:
        '201':
          description: Projet créé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '400':
          description: Données invalides
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /projects/{project_id}:
    get:
      tags:
        - Projects
      summary: Détails d'un projet
      description: Retourne les détails d'un projet spécifique
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Projet trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    put:
      tags:
        - Projects
      summary: Modifier un projet
      description: Met à jour les informations d'un projet
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectUpdate'
      responses:
        '200':
          description: Projet mis à jour
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    
    delete:
      tags:
        - Projects
      summary: Supprimer un projet
      description: Supprime définitivement un projet
      security:
        - bearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Projet supprimé
        '404':
          description: Projet non trouvé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Non autorisé
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  
  # ===========================================================================
  # Media Endpoints
  # ===========================================================================
...
security:
  - bearerAuth: []